Add header to SetTagProperties
