+++ src/modules/Preservation/views/AddScope/SetTagProperties/SetTagProperties.style.js
@@ -1,6 +1,15 @@
 import styled from 'styled-components';
 import { tokens } from '@equinor/eds-tokens';
 
+export const Header = styled.header`
+    display: flex;
+    align-items: baseline;
+    
+    h1 {
+        margin-right: calc(var(--grid-unit) * 2);
+    };
+`;
+
 export const InputContainer = styled.div`
     margin: calc(var(--grid-unit) * 2) 0px;
     display: flex;
@@ -11,6 +20,7 @@ export const InputContainer = styled.div`
 export const Container = styled.div`
     display: grid;
     grid-template-columns: 1fr auto;
+    margin-top: calc(var(--grid-unit) * 3);
 `;
 
 export const ButtonContainer = styled.div`
@@ -22,7 +32,6 @@ export const ButtonContainer = styled.div`
 export const FormFieldSpacer = styled.div`
     margin-left: calc(var(--grid-unit) * 2);
 `;
-
 
 export const ButtonContent = styled.span`
     display: flex;
