+++ src/modules/Preservation/views/AddScope/types.d.ts
@@ -1,14 +1,19 @@
 export interface Tag {
     tagNo: string;
+    description: string;
+    mcPkgNo: string;
 }
 
-export interface TagRow {
+export type TagRow = {
     tagNo: string;
     description: string;
     purchaseOrderNumber: string;
     commPkgNo: string;
     mcPkgNo: string;
     isPreserved: boolean;
+    tableData?: {
+        checked: boolean;
+    };
 }
 
 /**
